Refactor helm hook argument (#49)
* Refactor helm hook application to use a resource protocol

* Refactor apply_helm_hook function to use cdk8s.ApiObject and improve error handling for resource metadata

+++ ca_cdk8s_constructs/cdk8s_helm_chart.py
@@ -11,7 +11,6 @@ import yaml
 from aws_cdk import Duration
 from aws_cdk.aws_eks import HelmChart, ICluster
 from aws_cdk.aws_s3_assets import Asset
-from cdk8s_plus_32 import Job
 from constructs import Construct
 
 
@@ -181,29 +180,34 @@ class HelmHookDeletePolicy(Enum):
 
 
 def apply_helm_hook(
-    job: Job,
+    resource: cdk8s.ApiObject,
     hook_types: Sequence[HelmHookType],
     hook_delete_policy: Sequence[HelmHookDeletePolicy] = [
         HelmHookDeletePolicy.BEFORE_HOOK_CREATION
     ],
-    weight: int = None,
+    weight: int = 0,
 ) -> None:
-    """Apply helm hook annotations to a job.
+    """Apply helm hook annotations to a resource.
 
-    This applies the correct annotations to the job to ensure that the job is executed as
+    This applies the correct annotations to the resource to ensure that the resource is executed as
     a helm hook if the resource is deployed as a helm chart.
 
     Args:
-        job: The job to apply the helm hook to.
+        resource: The resource to apply the helm hook to. Must have a metadata attribute with add_annotation method.
         hook_types: The types of hook to apply.
         hook_delete_policy: The policy for deleting the hook resource after the hook has been executed.
-        weight: The weight of the hook.
+        weight: The weight specifies the order in which hooks are executed.
     """
-    job.metadata.add_annotation(
+
+    if not hasattr(resource, "metadata") or not hasattr(resource.metadata, "add_annotation"):
+        raise TypeError("Resource must have a metadata attribute with add_annotation method.")
+
+    resource.metadata.add_annotation(
         "helm.sh/hook-delete-policy", ",".join([policy.value for policy in hook_delete_policy])
     )
 
-    job.metadata.add_annotation("helm.sh/hook", ",".join([hook.value for hook in hook_types]))
+    resource.metadata.add_annotation(
+        "helm.sh/hook", ",".join([hook.value for hook in hook_types])
+    )
 
-    if weight:
-        job.metadata.add_annotation("helm.sh/hook-weight", str(weight))
+    resource.metadata.add_annotation("helm.sh/hook-weight", str(weight))
